Use images in europe north1
This is the first chacomit to use the new europe-north1 region for images.
This change is made to ensure that the images are stored in the correct region.
The change affects the GitHub workflows for building the chart and the image.
The new region is europe-north1, which is the designated region for the Dapla
lab Docker images.

This change is part of the ongoing effort to improve the Dapla lab Docker images
and to ensure that they are available in the correct region. The new region is
europe-north1, which is the designated region for the Dapla lab Docker images
and is expected to provide better performance and availability for the images.

The second commit is to update the GitHub workflows to use the new region for
the images. The workflows are updated to use the new region for both the chart
and the image. This ensures that the images are built and pushed to the correct
region, and that they are available for use in the Dapla lab.

Internal ref: [DPSKY-1047]